requirements.txt
flask_ml
onnxruntime
Pillow
numpy
